[Gold V] Title: 선 긋기, Time: 2568 ms, Memory: 319956 KB -BaekjoonHub

+++ 백준/Gold/2170. 선 긋기/선 긋기.java
@@ -33,7 +33,7 @@ public class Main {
                 pq.offer(line);
             }
         }
-        long total = 0;
+        int total = 0;
         while (!pq.isEmpty()) {
             Line line = pq.poll();
             total += line.end - line.start;
